+++ packages/babel-make-styles/__fixtures__/config-evaluation-rules/code.ts
@@ -1,13 +1,8 @@
 import { makeStyles } from '@fluentui/react-make-styles';
-
-const func = () => {
-  // This assignment has no sense, but it will prevent us from evaluation in AST
-  // This fixture uses "sampleEvaluator.js" in plugin's config so input we should get a different color
-  const color = 'red';
-
-  return { color };
-};
+import { colorRed } from './consts';
 
 export const useStyles = makeStyles({
-  root: func(),
+  // This import has no sense, but it will prevent us from evaluation in AST by Babel
+  // This fixture uses "sampleEvaluator.js" in plugin's config so input we should get a different color
+  root: { color: colorRed },
 });