+++ packages/babel-make-styles/src/plugin.ts
@@ -2,81 +2,38 @@ import { NodePath, PluginObj, PluginPass, types as t } from '@babel/core';
 import { declare } from '@babel/helper-plugin-utils';
 import { Module } from '@linaria/babel-preset';
 import shakerEvaluator from '@linaria/shaker';
-import { resolveStyleRulesForSlots, CSSRulesByBucket, StyleBucketName, MakeStylesStyle } from '@fluentui/make-styles';
+import { resolveStyleRulesForSlots, CSSRulesByBucket, StyleBucketName, GriffelStyle } from '@griffel/core';
 
 import { astify } from './utils/astify';
 import { evaluatePaths } from './utils/evaluatePaths';
-import { UNHANDLED_CASE_ERROR } from './constants';
 import { BabelPluginOptions } from './types';
 import { validateOptions } from './validateOptions';
-
-type AstStyleNode =
-  | { kind: 'PURE_OBJECT'; nodePath: NodePath<t.ObjectExpression> }
-  | {
-      kind: 'LAZY_OBJECT';
-      nodePath: NodePath<t.ObjectExpression>;
-      lazyPaths: NodePath<t.Expression | t.SpreadElement>[];
-    }
-  | { kind: 'LAZY_FUNCTION'; nodePath: NodePath<t.ArrowFunctionExpression | t.FunctionExpression> }
-  | { kind: 'LAZY_EXPRESSION_CALL'; nodePath: NodePath<t.CallExpression> }
-  | { kind: 'LAZY_SEQUENCE_EXPRESSION'; nodePath: NodePath<t.SequenceExpression> }
-  | { kind: 'LAZY_MEMBER'; nodePath: NodePath<t.MemberExpression> }
-  | { kind: 'LAZY_IDENTIFIER'; nodePath: NodePath<t.Identifier> }
-  | { kind: 'SPREAD'; nodePath: NodePath<t.SpreadElement>; spreadPath: NodePath<t.SpreadElement> };
 
 type BabelPluginState = PluginPass & {
   importDeclarationPath?: NodePath<t.ImportDeclaration>;
   requireDeclarationPath?: NodePath<t.VariableDeclarator>;
 
-  /** Contains all paths to calls of makeStyles(). */
+  definitionPaths?: NodePath<t.ObjectExpression>[];
   calleePaths?: NodePath<t.Identifier>[];
-
-  /** Contains AST nodes with that should be resolved. */
-  styleNodes?: AstStyleNode[];
 };
 
-/**
- * Parses a MemberExpression that defines token usage for extract parts that are needed to generate a CSS variable.
- *
- * @example
- *   getTokenParts(theme.alias.color.green.foreground1)
- *   // results in
- *   ['alias', 'color', 'green', 'foreground']
- */
-function getTokenParts(path: NodePath<t.MemberExpression>, result: string[] = []): string[] {
-  const objectPath = path.get('object');
-  const propertyPath = path.get('property');
+function getDefinitionPathFromMakeStylesCallExpression(
+  callExpression: NodePath<t.CallExpression>,
+): NodePath<t.ObjectExpression> {
+  const argumentPaths = callExpression.get('arguments') as NodePath<t.Node>[];
+  const hasValidArguments = Array.isArray(argumentPaths) && argumentPaths.length === 1;
 
-  if (objectPath.isMemberExpression()) {
-    getTokenParts(objectPath, result);
-  } else if (objectPath.isIdentifier()) {
-    // We ignore this part as it will return a first id (i.e. "theme")
-  } else {
-    throw objectPath.buildCodeFrameError(UNHANDLED_CASE_ERROR);
+  if (!hasValidArguments) {
+    throw new Error('makeStyles() function accepts only a single param');
   }
 
-  if (propertyPath.isIdentifier()) {
-    result.push(propertyPath.node.name);
+  const definitionsPath = argumentPaths[0];
+
+  if (!definitionsPath.isObjectExpression()) {
+    throw definitionsPath.buildCodeFrameError('makeStyles() function accepts only an object as a param');
   }
 
-  return result;
-}
-
-/**
- * Gets an Identifier from a MemberExpression that represents theme variable name.
- */
-function getMemberExpressionIdentifier(expressionPath: NodePath<t.MemberExpression>): NodePath<t.Identifier> {
-  const objectPath = expressionPath.get('object');
-
-  if (objectPath.isIdentifier()) {
-    return objectPath;
-  }
-
-  if (objectPath.isMemberExpression()) {
-    return getMemberExpressionIdentifier(objectPath);
-  }
-
-  throw objectPath.buildCodeFrameError(UNHANDLED_CASE_ERROR);
+  return definitionsPath;
 }
 
 /**
@@ -126,335 +83,6 @@ function isRequireDeclarator(path: NodePath<t.VariableDeclarator>): boolean {
   return false;
 }
 
-function namesToCssVariable(names: string[]): string {
-  return `var(--${names.join('-')})`;
-}
-
-/**
- * Processes an object (makeStyles argument) and collects smallest possible paths for evaluation later.
- */
-function processDefinitions(
-  argumentsPaths: NodePath<t.CallExpression['arguments'][number]>[],
-  state: BabelPluginState,
-): void {
-  const hasValidArguments = Array.isArray(argumentsPaths) && argumentsPaths.length === 1;
-
-  if (!hasValidArguments) {
-    throw new Error('makeStyles() function accepts only a single param');
-  }
-
-  const definitionsPath = argumentsPaths[0];
-
-  if (!definitionsPath.isObjectExpression()) {
-    throw definitionsPath.buildCodeFrameError('makeStyles() function accepts only an object as a param');
-  }
-
-  const styleSlots = definitionsPath.get('properties');
-
-  styleSlots.forEach((styleSlotPath: NodePath<t.SpreadElement | t.ObjectMethod | t.ObjectProperty>) => {
-    if (styleSlotPath.isObjectMethod()) {
-      throw styleSlotPath.buildCodeFrameError('Object methods are not supported for defining styles');
-    }
-
-    /**
-     * Needs lazy evaluation anyway.
-     *
-     * @example makeStyles({ ...SOME_STYLES })
-     */
-    if (styleSlotPath.isSpreadElement()) {
-      // Heads up!
-      //
-      // Babel can't evaluate spreads so we will fallback into VM execution anyway. Evaluation in VM replaces evaluated
-      // paths, this means that we will loose a "nodePath" as an identifier and will not be able to replace this path
-      // later.
-      // There we are cloning an existing spread and creating a wrapping spread:
-      // "{ ...fooBar }" becomes "{ ...{ ...forBar } }"
-      //                                 ^ this spread will be evaluated in VM
-      //                            ^ this spread will be used for replacement later
-      const spreadArgumentPath = styleSlotPath.get('argument');
-      const clonedArgumentNode = t.cloneNode(spreadArgumentPath.node);
-      const wrappingSpreadArgumentNode = t.objectExpression([t.spreadElement(clonedArgumentNode)]);
-
-      spreadArgumentPath.replaceWith(wrappingSpreadArgumentNode);
-      state.styleNodes?.push({
-        kind: 'SPREAD',
-        nodePath: styleSlotPath,
-        spreadPath: styleSlotPath.get('argument.properties.0') as NodePath<t.SpreadElement>,
-      });
-      return;
-    }
-
-    if (styleSlotPath.isObjectProperty()) {
-      const stylesPath = styleSlotPath.get('value');
-
-      /**
-       * Needs lazy evaluation anyway.
-       *
-       * @example makeStyles({ root: SOME_VARIABLE })
-       */
-      if (stylesPath.isIdentifier()) {
-        state.styleNodes?.push({
-          kind: 'LAZY_IDENTIFIER',
-          nodePath: stylesPath,
-        });
-        return;
-      }
-
-      /**
-       * May need lazy evaluation in less optimistic scenarios.
-       *
-       * @example
-       *    makeStyles({ root: { color: 'red' } }) // ✔ can be resolved in AST
-       *    makeStyles({ root: { color: SOME_VARIABLE } }) // ❌ lazy evaluation
-       *    makeStyles({ ...sharedStyles }) // ❌ lazy evaluation
-       */
-      if (stylesPath.isObjectExpression()) {
-        const propertiesPaths = stylesPath.get('properties');
-        const lazyPaths: NodePath<t.Expression | t.SpreadElement>[] = [];
-
-        propertiesPaths.forEach(propertyPath => {
-          if (propertyPath.isObjectMethod()) {
-            throw propertyPath.buildCodeFrameError('Object methods are not supported for defining styles');
-          }
-
-          if (propertyPath.isObjectProperty()) {
-            const keyPath = propertyPath.get('key');
-            const valuePath = propertyPath.get('value');
-
-            /**
-             * Computed properties may require lazy evaluation.
-             *
-             * @example
-             *    makeStyles({ [var]: { color: 'red' } })
-             *    makeStyles({ [`${var}`]: { color: SOME_VARIABLE } })
-             */
-            if (propertyPath.node.computed) {
-              lazyPaths.push(keyPath);
-            }
-
-            if (valuePath.isStringLiteral() || valuePath.isNullLiteral() || valuePath.isNumericLiteral()) {
-              return;
-            }
-
-            // TODO: properly support this case to avoid useless lazy evaluations (as they should be slower)
-            //       We should use recursive lookup there
-            //
-            // if (valuePath.isObjectExpression()) {
-            //   return;
-            // }
-
-            if (valuePath.isExpression()) {
-              lazyPaths.push(valuePath);
-              return;
-            }
-
-            throw valuePath.buildCodeFrameError(UNHANDLED_CASE_ERROR);
-          }
-
-          if (propertyPath.isSpreadElement()) {
-            lazyPaths.push(propertyPath);
-            return;
-          }
-
-          throw propertyPath.buildCodeFrameError(UNHANDLED_CASE_ERROR);
-        });
-
-        if (lazyPaths.length === 0) {
-          state.styleNodes?.push({
-            kind: 'PURE_OBJECT',
-            nodePath: stylesPath,
-          });
-          return;
-        }
-
-        state.styleNodes?.push({
-          kind: 'LAZY_OBJECT',
-          nodePath: stylesPath,
-          lazyPaths,
-        });
-        return;
-      }
-
-      /**
-       * A scenario when slots styles are represented by functions, in the worst case fallbacks to lazy
-       * evaluation.
-       *
-       * @example
-       *    // ✔ can be resolved in AST
-       *    makeStyles({ root: (t) => ({ color: t.red }) })
-       *    makeStyles({ root: () => ({ padding: '12px' }) })
-       *    // ❌ lazy evaluation
-       *    makeStyles({ root: (t) => ({ color: SOME_VARIABLE }) })
-       *    // ❌ lazy evaluation, the worst case as function contains body
-       *    makeStyles({ root: (t) => { return { color: SOME_VARIABLE } } })
-       */
-      if (stylesPath.isArrowFunctionExpression()) {
-        const paramError = 'A function in makeStyles() can only have a single param (for tokens/theme)';
-        if (stylesPath.get('params').length > 1) {
-          throw stylesPath.buildCodeFrameError(paramError);
-        }
-
-        const paramsPath = stylesPath.get('params.0') as NodePath<t.Node> | undefined;
-
-        if (paramsPath && !paramsPath.isIdentifier()) {
-          throw stylesPath.buildCodeFrameError(`${paramError} and it should be a valid identifier`);
-        }
-
-        const bodyPath = stylesPath.get('body');
-
-        /**
-         * Optimistic case, we may not need to use lazy evaluation 🚀
-         *
-         * @example
-         *    // ✔ can be resolved in AST
-         *    makeStyles({ root: (t) => ({ color: t.red }) })
-         *    makeStyles({ root: () => ({ padding: '12px' }) })
-         *    // ❌ lazy evaluation
-         *    makeStyles({ root: (t) => ({ color: SOME_VARIABLE }) })
-         */
-        if (bodyPath.isObjectExpression()) {
-          const propertiesPaths = bodyPath.get('properties');
-          const lazyPaths: NodePath<t.Expression | t.SpreadElement>[] = [];
-
-          propertiesPaths.forEach(propertyPath => {
-            if (propertyPath.isObjectMethod()) {
-              throw propertyPath.buildCodeFrameError('Object methods are not supported for defining styles');
-            }
-
-            if (propertyPath.isObjectProperty()) {
-              const valuePath = propertyPath.get('value');
-
-              if (valuePath.isStringLiteral() || valuePath.isNullLiteral() || valuePath.isNumericLiteral()) {
-                return;
-              }
-
-              // This condition resolves "theme.aliasColorGreenForeground1" to CSS variable
-              if (valuePath.isMemberExpression()) {
-                const identifierPath = getMemberExpressionIdentifier(valuePath);
-                const paramsName = paramsPath?.node.name;
-
-                if (paramsName && identifierPath.isIdentifier({ name: paramsName })) {
-                  const cssVariable = namesToCssVariable(getTokenParts(valuePath));
-
-                  valuePath.replaceWith(t.stringLiteral(cssVariable));
-                  return;
-                }
-
-                lazyPaths.push(valuePath);
-                return;
-              }
-
-              if (valuePath.isExpression()) {
-                lazyPaths.push(valuePath);
-                return;
-              }
-
-              throw valuePath.buildCodeFrameError(UNHANDLED_CASE_ERROR);
-            }
-
-            if (propertyPath.isSpreadElement()) {
-              lazyPaths.push(propertyPath);
-              return;
-            }
-
-            throw propertyPath.buildCodeFrameError(UNHANDLED_CASE_ERROR);
-          });
-
-          if (lazyPaths.length === 0) {
-            stylesPath.replaceWith(bodyPath);
-            state.styleNodes?.push({
-              kind: 'PURE_OBJECT',
-              // 👇 as we replaced an arrow function with its body, we can cast typings
-              nodePath: (stylesPath as unknown) as NodePath<t.ObjectExpression>,
-            });
-            return;
-          }
-
-          state.styleNodes?.push({
-            kind: 'LAZY_FUNCTION',
-            nodePath: stylesPath,
-          });
-          return;
-
-          // TODO: this can be lazy object once we will implement nested lookup, see object path for reference
-
-          // stylesPath.replaceWith(bodyPath);
-          // state.styleNodes?.push({
-          //   kind: 'LAZY_OBJECT',
-          //   // 👇 as we replaced an arrow function with its body, we can cast typings
-          //   nodePath: (stylesPath as unknown) as NodePath<t.ObjectExpression>,
-          //   lazyPaths,
-          // });
-          //
-          // return;
-        }
-
-        state.styleNodes?.push({
-          kind: 'LAZY_FUNCTION',
-          nodePath: stylesPath,
-        });
-        return;
-      }
-
-      /**
-       * A scenario when slots styles are represented by functions with body, fallbacks to lazy evaluation.
-       *
-       * @example
-       *    // ❌ lazy evaluation
-       *    makeStyles({ root: function (t) { return { color: SOME_VARIABLE } } })
-       */
-      if (stylesPath.isFunctionExpression()) {
-        state.styleNodes?.push({
-          kind: 'LAZY_FUNCTION',
-          nodePath: stylesPath,
-        });
-        return;
-      }
-
-      /**
-       * A scenario when slots styles are represented by a function call.
-       *
-       * @example
-       *    // ❌ lazy evaluation
-       *    makeStyles({ root: display() })
-       *    makeStyles({ root: typography.display() })
-       */
-      if (stylesPath.isCallExpression()) {
-        state.styleNodes?.push({ kind: 'LAZY_EXPRESSION_CALL', nodePath: stylesPath });
-        return;
-      }
-
-      /**
-       * A scenario when slots styles are represented by an object.
-       *
-       * @example
-       *    // ❌ lazy evaluation
-       *    makeStyles({ root: typography.display })
-       */
-      if (stylesPath.isMemberExpression()) {
-        state.styleNodes?.push({ kind: 'LAZY_MEMBER', nodePath: stylesPath });
-        return;
-      }
-
-      /**
-       * A scenario when slots styles are represented by a sequence expression.
-       *
-       * @example
-       *    // ❌ lazy evaluation
-       *    makeStyles({ root: (_a: { color: 'red' }, _a.backgroundColor: 'blue', _a ) })
-       *    makeStyles({ root: (_a: { color: SOME_VARIABLE }, _a.backgroundColor: 'blue', _a) })
-       *    makeStyles({ root: (_a: { color: 'red' }, _a[`. ${some_className}`]: { color: 'blue' }, _a) })
-       */
-      if (stylesPath.isSequenceExpression()) {
-        state.styleNodes?.push({ kind: 'LAZY_SEQUENCE_EXPRESSION', nodePath: stylesPath });
-        return;
-      }
-    }
-
-    throw styleSlotPath.buildCodeFrameError(UNHANDLED_CASE_ERROR);
-  });
-}
-
 /**
  * Rules that are returned by `resolveStyles()` are not deduplicated.
  * It's critical to filter out duplicates for build-time transform to avoid duplicated rules in a bundle.
@@ -497,8 +125,8 @@ export const plugin = declare<Partial<BabelPluginOptions>, PluginObj<BabelPlugin
     name: '@fluentui/babel-make-styles',
 
     pre() {
+      this.definitionPaths = [];
       this.calleePaths = [];
-      this.styleNodes = [];
     },
 
     visitor: {
@@ -513,85 +141,31 @@ export const plugin = declare<Partial<BabelPluginOptions>, PluginObj<BabelPlugin
             return;
           }
 
-          const pathsToEvaluate = state.styleNodes!.reduce<NodePath<t.Expression | t.SpreadElement>[]>(
-            (acc, styleNode) => {
-              if (
-                styleNode.kind === 'LAZY_IDENTIFIER' ||
-                styleNode.kind === 'LAZY_FUNCTION' ||
-                styleNode.kind === 'LAZY_MEMBER' ||
-                styleNode.kind === 'LAZY_EXPRESSION_CALL' ||
-                styleNode.kind === 'LAZY_SEQUENCE_EXPRESSION'
-              ) {
-                return [...acc, styleNode.nodePath];
-              }
+          if (state.definitionPaths) {
+            // Runs Babel AST processing or module evaluation for Node once for all arguments of makeStyles() calls once
+            evaluatePaths(path, state.file.opts.filename!, state.definitionPaths, pluginOptions);
 
-              if (styleNode.kind === 'LAZY_OBJECT') {
-                return [...acc, ...styleNode.lazyPaths];
-              }
-
-              if (styleNode.kind === 'PURE_OBJECT') {
-                return acc;
-              }
-
-              if (styleNode.kind === 'SPREAD') {
-                return [...acc, styleNode.spreadPath];
-              }
-
-              throw new Error(
-                `We don't support "styleNode.kind=${
-                  (styleNode as AstStyleNode).kind
-                }", this is a bug, please report it`,
-              );
-            },
-            [],
-          );
-
-          if (pathsToEvaluate.length > 0) {
-            evaluatePaths(path, state.file.opts.filename!, pathsToEvaluate, pluginOptions);
-          }
-
-          state.styleNodes?.forEach(styleNode => {
-            if (styleNode.kind === 'SPREAD') {
-              const nodePath = styleNode.nodePath;
-              const evaluationResult = (nodePath.get('argument') as NodePath<t.Expression>).evaluate();
+            state.definitionPaths.forEach(definitionPath => {
+              const callExpressionPath = definitionPath.findParent(parentPath =>
+                parentPath.isCallExpression(),
+              ) as NodePath<t.CallExpression>;
+              const evaluationResult = definitionPath.evaluate();
 
               if (!evaluationResult.confident) {
-                // This is undocumented but shows the specific statement that failed
                 // eslint-disable-next-line @typescript-eslint/no-explicit-any
                 const deoptPath = (evaluationResult as any).deopt as NodePath | undefined;
-                throw (deoptPath || nodePath).buildCodeFrameError(
+                throw (deoptPath || definitionPath).buildCodeFrameError(
                   'Evaluation of a code fragment failed, this is a bug, please report it',
                 );
               }
 
-              const stylesBySlots: Record<string /* slot*/, MakeStylesStyle> = evaluationResult.value;
+              const stylesBySlots: Record<string /* slot */, GriffelStyle> = evaluationResult.value;
+              const [classnamesMapping, cssRules] = resolveStyleRulesForSlots(stylesBySlots);
 
-              nodePath.replaceWithMultiple((astify(stylesBySlots) as t.ObjectExpression).properties);
-            }
-          });
-
-          state.calleePaths?.forEach(calleePath => {
-            const callExpressionPath = calleePath.findParent(parentPath =>
-              parentPath.isCallExpression(),
-            ) as NodePath<t.CallExpression>;
-            const argumentPath = callExpressionPath.get('arguments.0') as NodePath<t.ObjectExpression>;
-
-            const evaluationResult = argumentPath.evaluate();
-
-            if (!evaluationResult.confident) {
-              // eslint-disable-next-line @typescript-eslint/no-explicit-any
-              const deoptPath = (evaluationResult as any).deopt as NodePath | undefined;
-              throw (deoptPath || argumentPath).buildCodeFrameError(
-                'Evaluation of a code fragment failed, this is a bug, please report it',
-              );
-            }
-
-            const stylesBySlots: Record<string /* slot */, MakeStylesStyle> = evaluationResult.value;
-            const [classnamesMapping, cssRules] = resolveStyleRulesForSlots(stylesBySlots, 0);
-
-            // TODO: find a better way to replace arguments
-            callExpressionPath.node.arguments = [astify(classnamesMapping), astify(dedupeCSSRules(cssRules))];
-          });
+              // TODO: find a better way to replace arguments
+              callExpressionPath.node.arguments = [astify(classnamesMapping), astify(dedupeCSSRules(cssRules))];
+            });
+          }
 
           if (state.importDeclarationPath) {
             const specifiers = state.importDeclarationPath.get('specifiers');
@@ -617,8 +191,8 @@ export const plugin = declare<Partial<BabelPluginOptions>, PluginObj<BabelPlugin
             });
           }
 
-          if (state.calleePaths!.length > 0) {
-            state.calleePaths!.forEach(calleePath => {
+          if (state.calleePaths) {
+            state.calleePaths.forEach(calleePath => {
               calleePath.replaceWith(t.identifier('__styles'));
             });
           }
@@ -636,12 +210,16 @@ export const plugin = declare<Partial<BabelPluginOptions>, PluginObj<BabelPlugin
       VariableDeclarator(path, state) {
         if (isRequireDeclarator(path)) {
           state.requireDeclarationPath = path;
-          return;
         }
       },
 
       // eslint-disable-next-line @typescript-eslint/naming-convention
       CallExpression(path, state) {
+        /**
+         * Handles case when `makeStyles()` is `CallExpression`.
+         *
+         * @example makeStyles({})
+         */
         if (!state.importDeclarationPath) {
           return;
         }
@@ -652,12 +230,17 @@ export const plugin = declare<Partial<BabelPluginOptions>, PluginObj<BabelPlugin
           return;
         }
 
-        processDefinitions(path.get('arguments'), state);
+        state.definitionPaths!.push(getDefinitionPathFromMakeStylesCallExpression(path));
         state.calleePaths!.push(calleePath);
       },
 
       // eslint-disable-next-line @typescript-eslint/naming-convention
       MemberExpression(expressionPath, state) {
+        /**
+         * Handles case when `makeStyles()` is inside `MemberExpression`.
+         *
+         * @example module.makeStyles({})
+         */
         if (!state.requireDeclarationPath) {
           return;
         }
@@ -679,7 +262,7 @@ export const plugin = declare<Partial<BabelPluginOptions>, PluginObj<BabelPlugin
           return;
         }
 
-        processDefinitions(parentPath.get('arguments'), state);
+        state.definitionPaths!.push(getDefinitionPathFromMakeStylesCallExpression(parentPath));
         state.calleePaths!.push(propertyPath as NodePath<t.Identifier>);
       },
     },