+++ packages/react-badge/src/components/Badge/useBadgeStyles.ts
@@ -1,4 +1,4 @@
-import { shorthands, mergeClasses, makeStyles } from '@fluentui/react-make-styles';
+import { shorthands, mergeClasses, makeStyles } from '@griffel/react';
 import { tokens } from '@fluentui/react-theme';
 import type { BadgeState } from './Badge.types';
 
@@ -247,7 +247,7 @@ const useIconStyles = makeStyles({
 /**
  * Applies style classnames to slots
  */
-export const useBadgeStyles = (state: BadgeState): BadgeState => {
+export const useBadgeStyles_unstable = (state: BadgeState): BadgeState => {
   const rootStyles = useRootStyles();
 
   state.root.className = mergeClasses(