+++ packages/react-button/src/components/SplitButton/index.ts
@@ -2,4 +2,4 @@ export * from './SplitButton';
 export * from './SplitButton.types';
 export * from './renderSplitButton';
 export * from './useSplitButton';
-export { splitButtonClassName, useSplitButtonStyles } from './useSplitButtonStyles';
+export { splitButtonClassName, useSplitButtonStyles_unstable } from './useSplitButtonStyles';