+++ packages/react-dialog/src/components/Dialog/useDialogStyles.ts
@@ -17,7 +17,7 @@ const useStyles = makeStyles({
 /**
  * Apply styling to the Dialog slots based on the state
  */
-export const useDialogStyles = (state: DialogState): DialogState => {
+export const useDialogStyles_unstable = (state: DialogState): DialogState => {
   const styles = useStyles();
   state.root.className = mergeClasses(dialogClassName, styles.root, state.root.className);