+++ packages/react-divider/src/components/Divider/Divider.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { renderDivider } from './renderDivider';
-import { useDivider } from './useDivider';
-import { useDividerStyles } from './useDividerStyles';
+import { renderDivider_unstable } from './renderDivider';
+import { useDivider_unstable } from './useDivider';
+import { useDividerStyles_unstable } from './useDividerStyles';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
 import type { DividerProps } from './Divider.types';
 
@@ -9,11 +9,11 @@ import type { DividerProps } from './Divider.types';
  * A divider visually segments content into groups.
  */
 export const Divider: ForwardRefComponent<DividerProps> = React.forwardRef((props, ref) => {
-  const state = useDivider(props, ref);
+  const state = useDivider_unstable(props, ref);
 
-  useDividerStyles(state);
+  useDividerStyles_unstable(state);
 
-  return renderDivider(state);
+  return renderDivider_unstable(state);
 });
 
 Divider.displayName = 'Divider';