+++ packages/react-input/src/components/Input/renderInput.tsx
@@ -5,7 +5,7 @@ import type { InputSlots, InputState } from './Input.types';
 /**
  * Render the final JSX of Input
  */
-export const renderInput = (state: InputState) => {
+export const renderInput_unstable = (state: InputState) => {
   const { slots, slotProps } = getSlots<InputSlots>(state);
   return (
     <slots.root {...slotProps.root}>