+++ packages/react-portal/src/components/Portal/renderPortal.tsx
@@ -5,7 +5,7 @@ import type { PortalState } from './Portal.types';
 /**
  * Render the final JSX of Portal
  */
-export const renderPortal = (state: PortalState): React.ReactElement => {
+export const renderPortal_unstable = (state: PortalState): React.ReactElement => {
   return (
     <span hidden ref={state.virtualParentRootRef}>
       {state.shouldRender && state.mountNode && ReactDOM.createPortal(state.children, state.mountNode)}