+++ packages/react-radio/src/components/RadioItem/useRadioItem.ts
@@ -13,13 +13,13 @@ import type { RadioItemProps, RadioItemState } from './RadioItem.types';
 /**
  * Create the state required to render RadioItem.
  *
- * The returned state can be modified with hooks such as useRadioItemStyles,
- * before being passed to renderRadioItem.
+ * The returned state can be modified with hooks such as useRadioItemStyles_unstable,
+ * before being passed to renderRadioItem_unstable.
  *
  * @param props - props from this instance of RadioItem
  * @param ref - reference to root HTMLElement of RadioItem
  */
-export const useRadioItem = (props: RadioItemProps, ref: React.Ref<HTMLElement>): RadioItemState => {
+export const useRadioItem_unstable = (props: RadioItemProps, ref: React.Ref<HTMLElement>): RadioItemState => {
   const { id, labelPosition = 'inline' } = props;
   const [checked, setChecked] = useControllableState({
     defaultState: props.defaultChecked,