+++ packages/react-spinbutton/src/components/SpinButton/SpinButton.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { useSpinButton } from './useSpinButton';
-import { renderSpinButton } from './renderSpinButton';
-import { useSpinButtonStyles } from './useSpinButtonStyles';
+import { useSpinButton_unstable } from './useSpinButton';
+import { renderSpinButton_unstable } from './renderSpinButton';
+import { useSpinButtonStyles_unstable } from './useSpinButtonStyles';
 import type { SpinButtonProps } from './SpinButton.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
 
@@ -9,10 +9,10 @@ import type { ForwardRefComponent } from '@fluentui/react-utilities';
  * A SpinButton allows someone to incrementally adjust a value in small steps.
  */
 export const SpinButton: ForwardRefComponent<SpinButtonProps> = React.forwardRef((props, ref) => {
-  const state = useSpinButton(props, ref);
+  const state = useSpinButton_unstable(props, ref);
 
-  useSpinButtonStyles(state);
-  return renderSpinButton(state);
+  useSpinButtonStyles_unstable(state);
+  return renderSpinButton_unstable(state);
 });
 
 SpinButton.displayName = 'SpinButton';