+++ packages/react-switch/src/components/Switch/Switch.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { useSwitch } from './useSwitch';
-import { renderSwitch } from './renderSwitch';
-import { useSwitchStyles } from './useSwitchStyles';
+import { useSwitch_unstable } from './useSwitch';
+import { renderSwitch_unstable } from './renderSwitch';
+import { useSwitchStyles_unstable } from './useSwitchStyles';
 import type { SwitchProps } from './Switch.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
 
@@ -9,11 +9,11 @@ import type { ForwardRefComponent } from '@fluentui/react-utilities';
  * The Switch control enables users to trigger an option on or off through pressing on the component.
  */
 export const Switch: ForwardRefComponent<SwitchProps> = React.forwardRef((props, ref) => {
-  const state = useSwitch(props, ref);
+  const state = useSwitch_unstable(props, ref);
 
-  useSwitchStyles(state);
+  useSwitchStyles_unstable(state);
 
-  return renderSwitch(state);
+  return renderSwitch_unstable(state);
 });
 
 Switch.displayName = 'Switch';