+++ packages/react-monaco-editor/CHANGELOG.md
@@ -1,8 +1,62 @@
 # Change Log - @fluentui/react-monaco-editor
 
-This log was last generated on Wed, 12 Jan 2022 07:32:29 GMT and should not be manually modified.
+This log was last generated on Tue, 25 Jan 2022 07:30:06 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.2.69](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.2.69)
+
+Tue, 25 Jan 2022 07:30:06 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.2.68..@fluentui/react-monaco-editor_v1.2.69)
+
+### Patches
+
+- Bump @fluentui/react to v8.51.0 ([PR #20805](https://github.com/microsoft/fluentui/pull/20805) by beachball)
+
+## [1.2.68](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.2.68)
+
+Mon, 24 Jan 2022 07:26:13 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.2.67..@fluentui/react-monaco-editor_v1.2.68)
+
+### Patches
+
+- Bump @fluentui/react to v8.50.1 ([PR #21362](https://github.com/microsoft/fluentui/pull/21362) by beachball)
+
+## [1.2.67](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.2.67)
+
+Fri, 21 Jan 2022 07:26:06 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.2.66..@fluentui/react-monaco-editor_v1.2.67)
+
+### Patches
+
+- Bump @fluentui/react to v8.50.0 ([PR #21280](https://github.com/microsoft/fluentui/pull/21280) by beachball)
+
+## [1.2.66](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.2.66)
+
+Mon, 17 Jan 2022 09:43:48 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.2.65..@fluentui/react-monaco-editor_v1.2.66)
+
+### Patches
+
+- Bump @fluentui/react to v8.49.7 ([PR #21274](https://github.com/microsoft/fluentui/pull/21274) by beachball)
+
+## [1.2.65](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.2.65)
+
+Fri, 14 Jan 2022 07:28:39 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.2.64..@fluentui/react-monaco-editor_v1.2.65)
+
+### Patches
+
+- Bump @fluentui/react to v8.49.6 ([PR #21215](https://github.com/microsoft/fluentui/pull/21215) by beachball)
+
+## [1.2.64](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.2.64)
+
+Thu, 13 Jan 2022 07:30:58 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-monaco-editor_v1.2.63..@fluentui/react-monaco-editor_v1.2.64)
+
+### Patches
+
+- Bump @fluentui/react to v8.49.5 ([PR #21196](https://github.com/microsoft/fluentui/pull/21196) by beachball)
 
 ## [1.2.63](https://github.com/microsoft/fluentui/tree/@fluentui/react-monaco-editor_v1.2.63)